Executor logic separated from Poolmgr backend completely, placeholder for new backend
