+++ charts/fission-core/templates/deployment.yaml
@@ -226,16 +226,16 @@ spec:
         image: "{{ .Values.image }}:{{ .Values.imageTag }}"
         imagePullPolicy: {{ .Values.pullPolicy }}
         command: ["/fission-bundle"]
-        args: ["--routerPort", "8888", "--poolmgrUrl", "http://poolmgr.{{ .Release.Namespace }}"]
+        args: ["--routerPort", "8888", "--executorUrl", "http://executor.{{ .Release.Namespace }}"]
       serviceAccount: fission-svc
 
 ---
 apiVersion: v1
 kind: Service
 metadata:
-  name: poolmgr
+  name: executor
   labels:
-    svc: poolmgr
+    svc: executor
     chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
 spec:
   type: ClusterIP
@@ -243,13 +243,13 @@ spec:
   - port: 80
     targetPort: 8888
   selector:
-    svc: poolmgr
+    svc: executor
 
 ---
 apiVersion: extensions/v1beta1
 kind: Deployment
 metadata:
-  name: poolmgr
+  name: executor
   labels:
     chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
 spec:
@@ -257,14 +257,14 @@ spec:
   template:
     metadata:
       labels:
-        svc: poolmgr
+        svc: executor
     spec:
       containers:
-      - name: poolmgr
+      - name: executor
         image: "{{ .Values.image }}:{{ .Values.imageTag }}"
         imagePullPolicy: {{ .Values.pullPolicy }}
         command: ["/fission-bundle"]
-        args: ["--poolmgrPort", "8888", "--namespace", "{{ .Values.functionNamespace }}", "--fission-namespace", "{{ .Release.Namespace }}"]
+        args: ["--executorPort", "8888", "--namespace", "{{ .Values.functionNamespace }}", "--fission-namespace", "{{ .Release.Namespace }}"]
         env:
         - name: FETCHER_IMAGE
           value: "{{ .Values.fetcherImage }}:{{ .Values.fetcherImageTag }}"
